Fix external resources tests
